+++ packages/core/pluggableElementTypes/ViewType.ts
@@ -9,10 +9,10 @@ type BasicView = React.ComponentType<{
   model: any
   session?: IAnyStateTreeNode
 }>
-type ViewReactComponent = React.LazyExoticComponent<BasicView> | BasicView
+type ViewComponentType = React.LazyExoticComponent<BasicView> | BasicView
 
 export default class ViewType extends PluggableElementBase {
-  ReactComponent: ViewReactComponent
+  ReactComponent: ViewComponentType
 
   stateModel: IAnyModelType
 
@@ -29,7 +29,7 @@ export default class ViewType extends PluggableElementBase {
   constructor(stuff: {
     name: string
     displayName?: string
-    ReactComponent: ViewReactComponent
+    ReactComponent: ViewComponentType
     stateModel: IAnyModelType
     extendedName?: string
   }) {
