--- packages/core/pluggableElementTypes/renderers/CircularChordRendererType.ts
@@ -1,3 +1,0 @@
-import FeatureRenderer from './FeatureRendererType'
-
-export default class CircularChordRendererType extends FeatureRenderer {}
