+++ packages/core/pluggableElementTypes/renderers/ServerSideRenderedContent.tsx
@@ -1,26 +1,23 @@
 import React, { useEffect, useRef } from 'react'
 import { ThemeProvider } from '@mui/material/styles'
 import { hydrate, unmountComponentAtNode } from 'react-dom'
+
+// locals
 import { createJBrowseTheme } from '../../ui'
 import { rIC } from '../../util'
 import { ResultsSerialized, RenderArgs } from './ServerSideRendererType'
 
-interface ServerSideRenderedContentProps extends ResultsSerialized, RenderArgs {
+interface Props extends ResultsSerialized, RenderArgs {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   RenderingComponent: React.ComponentType<any>
 }
 
-export default function ServerSideRenderedContent(
-  props: ServerSideRenderedContentProps,
-) {
-  const ssrContainerNode = useRef<HTMLDivElement>(null)
-
-  const { theme, html, RenderingComponent, ...rest } = props
-
+export default function ({ theme, html, RenderingComponent, ...rest }: Props) {
+  const ref = useRef<HTMLDivElement>(null)
   const jbrowseTheme = createJBrowseTheme(theme)
 
   useEffect(() => {
-    const domNode = ssrContainerNode.current
+    const domNode = ref.current
     function doHydrate() {
       if (domNode) {
         if (domNode) {
@@ -58,11 +55,5 @@ export default function ServerSideRenderedContent(
     }
   }, [html, jbrowseTheme, rest, RenderingComponent])
 
-  return (
-    <div
-      ref={ssrContainerNode}
-      data-html-size={html.length}
-      className="ssr-container"
-    />
-  )
+  return <div ref={ref} />
 }
