+++ packages/core/rpc/coreRpcMethods.ts
@@ -1,372 +1,22 @@
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
-import { toArray } from 'rxjs/operators'
-import {
-  freeAdapterResources,
-  getAdapter,
-} from '../data_adapters/dataAdapterCache'
-import RpcMethodType from '../pluggableElementTypes/RpcMethodType'
-import ServerSideRendererType, {
-  RenderArgs as ServerSideRenderArgs,
-  RenderArgsSerialized as ServerSideRenderArgsSerialized,
-  RenderResults,
-  ResultsSerialized,
-} from '../pluggableElementTypes/renderers/ServerSideRendererType'
-import { RemoteAbortSignal } from './remoteAbortSignals'
-import { isFeatureAdapter } from '../data_adapters/BaseAdapter'
-import {
-  checkAbortSignal,
-  renameRegionsIfNeeded,
-  getLayoutId,
-  Region,
-} from '../util'
-import SimpleFeature, { SimpleFeatureSerialized } from '../util/simpleFeature'
+import CoreGetRefNames from './methods/CoreGetRefNames'
+import CoreGetMetadata from './methods/CoreGetMetadata'
+import CoreGetFeatureDetails from './methods/CoreGetFeatureDetails'
+import CoreGetFeatures from './methods/CoreGetFeatures'
+import CoreGetFileInfo from './methods/CoreGetFileInfo'
+import CoreFreeResources from './methods/CoreFreeResources'
+import CoreRender from './methods/CoreRender'
+import CoreEstimateRegionStats from './methods/CoreEstimateRegionStats'
+import { RenderArgs } from './methods/util'
 
-export class CoreGetRefNames extends RpcMethodType {
-  name = 'CoreGetRefNames'
+export type { RenderArgs }
 
-  async execute(
-    args: {
-      sessionId: string
-      signal: RemoteAbortSignal
-      adapterConfig: {}
-    },
-    rpcDriver: string,
-  ) {
-    const pm = this.pluginManager
-    const deserializedArgs = await this.deserializeArguments(args, rpcDriver)
-    const { sessionId, adapterConfig } = deserializedArgs
-    const { dataAdapter } = await getAdapter(pm, sessionId, adapterConfig)
-
-    if (isFeatureAdapter(dataAdapter)) {
-      return dataAdapter.getRefNames(deserializedArgs)
-    }
-    return []
-  }
-}
-export class CoreGetFileInfo extends RpcMethodType {
-  name = 'CoreGetInfo'
-
-  async execute(
-    args: {
-      sessionId: string
-      signal: RemoteAbortSignal
-      adapterConfig: {}
-    },
-    rpcDriver: string,
-  ) {
-    const pm = this.pluginManager
-    const deserializedArgs = await this.deserializeArguments(args, rpcDriver)
-    const { sessionId, adapterConfig } = deserializedArgs
-    const { dataAdapter } = await getAdapter(pm, sessionId, adapterConfig)
-    return isFeatureAdapter(dataAdapter)
-      ? dataAdapter.getHeader(deserializedArgs)
-      : null
-  }
-}
-
-export class CoreGetMetadata extends RpcMethodType {
-  name = 'CoreGetMetadata'
-
-  async execute(
-    args: {
-      sessionId: string
-      signal: RemoteAbortSignal
-      adapterConfig: {}
-    },
-    rpcDriver: string,
-  ) {
-    const pm = this.pluginManager
-    const deserializedArgs = await this.deserializeArguments(args, rpcDriver)
-    const { sessionId, adapterConfig } = deserializedArgs
-    const { dataAdapter } = await getAdapter(pm, sessionId, adapterConfig)
-    return isFeatureAdapter(dataAdapter)
-      ? dataAdapter.getMetadata(deserializedArgs)
-      : null
-  }
-}
-
-export class CoreGetFeatures extends RpcMethodType {
-  name = 'CoreGetFeatures'
-
-  async deserializeReturn(
-    feats: SimpleFeatureSerialized[],
-    args: unknown,
-    rpcDriver: string,
-  ) {
-    const superDeserialized = (await super.deserializeReturn(
-      feats,
-      args,
-      rpcDriver,
-    )) as SimpleFeatureSerialized[]
-    return superDeserialized.map(feat => new SimpleFeature(feat))
-  }
-
-  async serializeArguments(args: RenderArgs, rpcDriver: string) {
-    const { rootModel } = this.pluginManager
-    const assemblyManager = rootModel!.session!.assemblyManager
-    const renamedArgs = await renameRegionsIfNeeded(assemblyManager, args)
-    return super.serializeArguments(
-      renamedArgs,
-      rpcDriver,
-    ) as Promise<RenderArgs>
-  }
-
-  async execute(
-    args: {
-      sessionId: string
-      regions: Region[]
-      adapterConfig: {}
-      signal?: RemoteAbortSignal
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      opts?: any
-    },
-    rpcDriver: string,
-  ) {
-    const pm = this.pluginManager
-    const deserializedArgs = await this.deserializeArguments(args, rpcDriver)
-    const { signal, sessionId, adapterConfig, regions, opts } = deserializedArgs
-    const { dataAdapter } = await getAdapter(pm, sessionId, adapterConfig)
-    if (!isFeatureAdapter(dataAdapter)) {
-      throw new Error('Adapter does not support retrieving features')
-    }
-    const ret = dataAdapter.getFeaturesInMultipleRegions(regions, {
-      ...opts,
-      signal,
-    })
-    const r = await ret.pipe(toArray()).toPromise()
-    return r.map(f => f.toJSON())
-  }
-}
-
-/**
- * free up any resources (e.g. cached adapter objects)
- * that are only associated with the given track ID.
- *
- * returns number of objects deleted
- */
-export class CoreFreeResources extends RpcMethodType {
-  name = 'CoreFreeResources'
-
-  async execute(specification: {}) {
-    let deleteCount = 0
-
-    deleteCount += freeAdapterResources(specification)
-
-    // pass the freeResources hint along to all the renderers as well
-    this.pluginManager.getRendererTypes().forEach(renderer => {
-      const count = renderer.freeResources(/* specification */)
-      if (count) {
-        deleteCount += count
-      }
-    })
-
-    return deleteCount
-  }
-  async serializeArguments(args: {}, _rpcDriver: string): Promise<{}> {
-    return args
-  }
-}
-
-export interface RenderArgs extends ServerSideRenderArgs {
-  adapterConfig: {}
-  rendererType: string
-}
-
-export interface RenderArgsSerialized extends ServerSideRenderArgsSerialized {
-  assemblyName: string
-  regions: Region[]
-  adapterConfig: {}
-  rendererType: string
-}
-
-export class CoreEstimateRegionStats extends RpcMethodType {
-  name = 'CoreEstimateRegionStats'
-
-  async serializeArguments(
-    args: RenderArgs & {
-      signal?: AbortSignal
-      statusCallback?: (arg: string) => void
-    },
-    rpcDriver: string,
-  ) {
-    const { rootModel } = this.pluginManager
-    const assemblyManager = rootModel!.session!.assemblyManager
-    const renamedArgs = await renameRegionsIfNeeded(assemblyManager, {
-      ...args,
-      filters: args.filters?.toJSON().filters,
-    })
-
-    return super.serializeArguments(renamedArgs, rpcDriver)
-  }
-
-  async execute(
-    args: {
-      adapterConfig: {}
-      regions: Region[]
-      signal?: RemoteAbortSignal
-      headers?: Record<string, string>
-      sessionId: string
-    },
-    rpcDriver: string,
-  ) {
-    const pm = this.pluginManager
-    const deserializedArgs = await this.deserializeArguments(args, rpcDriver)
-    const { adapterConfig, sessionId, regions } = deserializedArgs
-    const { dataAdapter } = await getAdapter(pm, sessionId, adapterConfig)
-
-    if (!isFeatureAdapter(dataAdapter)) {
-      throw new Error('Adapter does not support retrieving features')
-    }
-    return dataAdapter.estimateRegionsStats(regions, deserializedArgs)
-  }
-}
-
-/**
- * fetches features from an adapter and call a renderer with them
- */
-export class CoreRender extends RpcMethodType {
-  name = 'CoreRender'
-
-  async serializeArguments(args: RenderArgs, rpcDriver: string) {
-    const { rootModel } = this.pluginManager
-    const assemblyManager = rootModel!.session!.assemblyManager
-    const renamedArgs = await renameRegionsIfNeeded(assemblyManager, args)
-
-    const superArgs = (await super.serializeArguments(
-      renamedArgs,
-      rpcDriver,
-    )) as RenderArgs
-    if (rpcDriver === 'MainThreadRpcDriver') {
-      return superArgs
-    }
-
-    const { rendererType } = args
-
-    const RendererType = validateRendererType(
-      rendererType,
-      this.pluginManager.getRendererType(rendererType),
-    )
-
-    return RendererType.serializeArgsInClient(superArgs)
-  }
-
-  async execute(
-    args: RenderArgsSerialized & { signal?: RemoteAbortSignal },
-    rpcDriver: string,
-  ) {
-    let deserializedArgs = args
-    if (rpcDriver !== 'MainThreadRpcDriver') {
-      deserializedArgs = await this.deserializeArguments(args, rpcDriver)
-    }
-    const { sessionId, rendererType, signal } = deserializedArgs
-    if (!sessionId) {
-      throw new Error('must pass a unique session id')
-    }
-
-    checkAbortSignal(signal)
-
-    const RendererType = validateRendererType(
-      rendererType,
-      this.pluginManager.getRendererType(rendererType),
-    )
-
-    const result =
-      rpcDriver === 'MainThreadRpcDriver'
-        ? await RendererType.render(deserializedArgs)
-        : await RendererType.renderInWorker(deserializedArgs)
-
-    checkAbortSignal(signal)
-    return result
-  }
-
-  async deserializeReturn(
-    serializedReturn: RenderResults | ResultsSerialized,
-    args: RenderArgs,
-    rpcDriver: string,
-  ): Promise<unknown> {
-    const des = await super.deserializeReturn(serializedReturn, args, rpcDriver)
-    if (rpcDriver === 'MainThreadRpcDriver') {
-      return des
-    }
-
-    const { rendererType } = args
-    const RendererType = validateRendererType(
-      rendererType,
-      this.pluginManager.getRendererType(rendererType),
-    )
-    return RendererType.deserializeResultsInClient(
-      des as ResultsSerialized,
-      args,
-    )
-  }
-}
-
-/**
- * fetches features from an adapter and call a renderer with them
- */
-export class CoreGetFeatureDetails extends RpcMethodType {
-  name = 'CoreGetFeatureDetails'
-
-  async serializeArguments(args: RenderArgs, rpcDriver: string) {
-    const { rootModel } = this.pluginManager
-    const assemblyManager = rootModel!.session!.assemblyManager
-    const renamedArgs = await renameRegionsIfNeeded(assemblyManager, args)
-    const superArgs = (await super.serializeArguments(
-      renamedArgs,
-      rpcDriver,
-    )) as RenderArgs
-    if (rpcDriver === 'MainThreadRpcDriver') {
-      return superArgs
-    }
-
-    const { rendererType } = args
-
-    const RendererType = validateRendererType(
-      rendererType,
-      this.pluginManager.getRendererType(rendererType),
-    )
-
-    return RendererType.serializeArgsInClient(superArgs)
-  }
-
-  async execute(
-    args: RenderArgsSerialized & { signal?: RemoteAbortSignal },
-    rpcDriver: string,
-  ) {
-    let deserializedArgs = args
-    if (rpcDriver !== 'MainThreadRpcDriver') {
-      deserializedArgs = await this.deserializeArguments(args, rpcDriver)
-    }
-    const { rendererType, featureId } = deserializedArgs
-    const RendererType = validateRendererType(
-      rendererType,
-      this.pluginManager.getRendererType(rendererType),
-    )
-
-    // @ts-ignore
-    const sess = RendererType.sessions[getLayoutId(args)]
-    const { layout } = sess.cachedLayout
-    const xref = layout.getDataByID(featureId)
-
-    return { feature: xref.toJSON() }
-  }
-}
-
-function validateRendererType<T>(rendererType: string, RendererType: T) {
-  if (!RendererType) {
-    throw new Error(`renderer "${rendererType}" not found`)
-  }
-  // @ts-ignore
-  if (!RendererType.ReactComponent) {
-    throw new Error(
-      `renderer ${rendererType} has no ReactComponent, it may not be completely implemented yet`,
-    )
-  }
-
-  if (!(RendererType instanceof ServerSideRendererType)) {
-    throw new Error(
-      'CoreRender requires a renderer that is a subclass of ServerSideRendererType',
-    )
-  }
-  return RendererType
+export {
+  CoreGetRefNames,
+  CoreGetFileInfo,
+  CoreGetMetadata,
+  CoreGetFeatures,
+  CoreRender,
+  CoreGetFeatureDetails,
+  CoreFreeResources,
+  CoreEstimateRegionStats,
 }
