+++ packages/core/rpc/methods/CoreGetFileInfo.ts
@@ -1,0 +1,26 @@
+import { getAdapter } from '../../data_adapters/dataAdapterCache'
+import RpcMethodType from '../../pluggableElementTypes/RpcMethodType'
+
+import { RemoteAbortSignal } from '../remoteAbortSignals'
+import { isFeatureAdapter } from '../../data_adapters/BaseAdapter'
+
+export default class CoreGetFileInfo extends RpcMethodType {
+  name = 'CoreGetInfo'
+
+  async execute(
+    args: {
+      sessionId: string
+      signal: RemoteAbortSignal
+      adapterConfig: {}
+    },
+    rpcDriver: string,
+  ) {
+    const pm = this.pluginManager
+    const deserializedArgs = await this.deserializeArguments(args, rpcDriver)
+    const { sessionId, adapterConfig } = deserializedArgs
+    const { dataAdapter } = await getAdapter(pm, sessionId, adapterConfig)
+    return isFeatureAdapter(dataAdapter)
+      ? dataAdapter.getHeader(deserializedArgs)
+      : null
+  }
+}
