+++ plugins/circular-view/package.json
@@ -37,7 +37,8 @@
     "clean": "rimraf dist esm *.tsbuildinfo"
   },
   "dependencies": {
-    "@mui/icons-material": "^5.0.1"
+    "@mui/icons-material": "^5.0.1",
+    "clone": "^2.1.2"
   },
   "peerDependencies": {
     "@jbrowse/core": "^2.0.0",
