+++ plugins/circular-view/src/BaseChordDisplay/components/BaseChordDisplay.tsx
@@ -1,16 +1,26 @@
 import React from 'react'
 import { observer } from 'mobx-react'
+
+// locals
 import Loading from './Loading'
 import DisplayError from './DisplayError'
-import RpcRenderedSvgGroup from './RpcRenderedSvgGroup'
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function BaseChordDisplay({ display }: any) {
+
+export default observer(function BaseChordDisplay({
+  display,
+}: {
+  display: {
+    filled: boolean
+    error: unknown
+    reactElement: React.ReactElement
+    renderProps: { radius: number }
+  }
+}) {
   if (display.error) {
     return <DisplayError model={display} />
   }
   if (!display.filled) {
     return <Loading model={display} />
   }
-  return <RpcRenderedSvgGroup model={display} />
-}
-export default observer(BaseChordDisplay)
+
+  return display.reactElement
+})
