+++ plugins/circular-view/src/BaseChordDisplay/components/Loading.tsx
@@ -48,66 +48,63 @@ const useStyles = makeStyles()(theme => {
   }
 })
 
-// 'repeating-linear-gradient(45deg, transparent, transparent 5px, rgba(255,255,255,.5) 5px, rgba(255,255,255,.5) 10px)',
-const Loading = observer(
-  ({
-    model: {
-      renderProps: { radius },
-    },
-  }: {
-    model: { renderProps: { radius: number } }
-  }) => {
-    const { classes } = useStyles()
-
-    // only show the loading message after 400ms to prevent excessive flickering
-    const [shown, setShown] = useState(false)
-    useEffect(() => {
-      const timeout = setTimeout(() => setShown(true), 400)
-      return () => clearTimeout(timeout)
-    })
-
-    return !shown ? null : (
-      <g>
-        <defs>
-          <pattern
-            id="diagonalHatch"
-            width="10"
-            height="10"
-            patternTransform="rotate(45 0 0)"
-            patternUnits="userSpaceOnUse"
-          >
-            <line
-              x1="0"
-              y1="0"
-              x2="0"
-              y2="10"
-              style={{ stroke: 'rgba(255,255,255,0.5)', strokeWidth: 10 }}
-            />
-          </pattern>
-        </defs>
-        <circle cx="0" cy="0" r={radius} fill="#f1f1f1" />
-        <circle cx="0" cy="0" r={radius} fill="url(#diagonalHatch)" />
-        <text
-          x="0"
-          y="0"
-          transform="rotate(90 0 0)"
-          dominantBaseline="middle"
-          textAnchor="middle"
-        >
-          Loading&hellip;
-        </text>
-        <circle
-          className={classes.path}
-          fill="none"
-          strokeWidth="4"
-          strokeLinecap="round"
-          cx="0"
-          cy="0"
-          r="60"
-        />
-      </g>
-    )
+const Loading = observer(function ({
+  model: {
+    renderProps: { radius },
   },
-)
+}: {
+  model: { renderProps: { radius: number } }
+}) {
+  const { classes } = useStyles()
+
+  // only show the loading message after 400ms to prevent excessive flickering
+  const [shown, setShown] = useState(false)
+  useEffect(() => {
+    const timeout = setTimeout(() => setShown(true), 400)
+    return () => clearTimeout(timeout)
+  })
+
+  return !shown ? null : (
+    <g>
+      <defs>
+        <pattern
+          id="diagonalHatch"
+          width="10"
+          height="10"
+          patternTransform="rotate(45 0 0)"
+          patternUnits="userSpaceOnUse"
+        >
+          <line
+            x1="0"
+            y1="0"
+            x2="0"
+            y2="10"
+            style={{ stroke: 'rgba(255,255,255,0.5)', strokeWidth: 10 }}
+          />
+        </pattern>
+      </defs>
+      <circle cx="0" cy="0" r={radius} fill="#f1f1f1" />
+      <circle cx="0" cy="0" r={radius} fill="url(#diagonalHatch)" />
+      <text
+        x="0"
+        y="0"
+        transform="rotate(90 0 0)"
+        dominantBaseline="middle"
+        textAnchor="middle"
+      >
+        Loading&hellip;
+      </text>
+      <circle
+        className={classes.path}
+        fill="none"
+        strokeWidth="4"
+        strokeLinecap="round"
+        cx="0"
+        cy="0"
+        r="60"
+      />
+    </g>
+  )
+})
 
 export default Loading
