+++ plugins/variants/src/StructuralVariantChordRenderer/index.ts
@@ -3,16 +3,15 @@ import PluginManager from '@jbrowse/core/PluginManager'
 import configSchema from './configSchema'
 import ReactComponent from './ReactComponent'
 
-const ChordRendererConfigF = (pluginManager: PluginManager) => {
+export default (pluginManager: PluginManager) => {
   pluginManager.addRendererType(
     () =>
       new ChordRendererType({
         name: 'StructuralVariantChordRenderer',
+        displayName: 'SV chord renderer',
         ReactComponent,
         configSchema,
         pluginManager,
       }),
   )
 }
-
-export default ChordRendererConfigF
